Fixing bug in the creation of the counts map
